+++ backend-shared/CMakeLists.txt
@@ -1,2 +1,9 @@
 # backend functionality shared between Desktop (UI) and Mobile (QML)
 
+set(BACKEND_SRCS
+	exportfuncs.cpp
+	exportfuncs.h
+)
+
+add_library(subsurface_backend_shared STATIC ${BACKEND_SRCS})
+target_link_libraries(subsurface_backend_shared ${QT_LIBRARIES})
