+++ backend-shared/exportfuncs.cpp
