+++ backend-shared/exportfuncs.h
